REGRESSION (r273154): Speedometer 2 is 1-2% regressed on iOS https://bugs.webkit.org/show_bug.cgi?id=222663 <rdar://problem/74715929>
Reviewed by Ryosuke Niwa.

A 1-2% regression was observed in Speedometer 2 after enabling the iOS
Form Control Refresh (FCR). Traces indicated an increased amount of time
spent in style resolution. Note that currently the FCR styles are
appended to the base UA stylesheet at runtime, to apply the necessary
changes for the new appearance, while making it possible to toggle the
feature.

To fix, the FCR stylesheet is merged into the base UA stylesheet. This
avoids ruleset duplication, which made it more expensive to resolve
styles. To preserve the old appearance when the feature is runtime
disabled, the old styles are moved into a separate stylesheet, appended
at runtime.

The fix was verified as an equivalent progression using internal A/B
testing infrastructure.

Note that this solution means that disabling the FCR will result in a
Speedometer regression, due to ruleset duplication. However, this should
not be a concern since the feature is controlled by an internal switch,
is enabled by default, and is not intended to be disabled.

* css/formControlsIOS.css:
* css/html.css:
* style/UserAgentStyle.cpp:
(WebCore::Style::UserAgentStyle::ensureDefaultStyleSheetsForElement):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273839 268f45cc-cd09-0410-ab3c-d52691b4dbfc