[WTF] Add System.framework PrivateHeaders to header path of WTF https://bugs.webkit.org/show_bug.cgi?id=222688
Reviewed by Tim Horton.

The other non-third-party components excluding WTF have System.framework PrivateHeaders in their header search path.
This causes the problem that we cannot use some system headers (e.g. `<machine/cpu_capabilities.h>`) in WTF while it is
usable in JavaScriptCore and the other components. This patch adds it as the other components did.

* Configurations/Base.xcconfig:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273849 268f45cc-cd09-0410-ab3c-d52691b4dbfc