[macOS][WK2] Changing the system language does not update navigator.language https://bugs.webkit.org/show_bug.cgi?id=222619
Reviewed by Per Arne Vollan.

Source/WebKit:

* Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceMain.mm:
(WebKit::setAppleLanguagesPreference):
- When ENABLE(CFPREFS_DIRECT_MODE), we don't need to set the AppleLanguages
in a volatile domain. This does not change behavior but I made this change
so that I don't need to set AppleLanguages in the volatile domain in
WebProcessCocoa when the preference changes.
- Also fix a memory leak for newArguments.

* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess):
If CFPREFS_DIRECT_MODE is not enabled, we need to listen for AppleLanguagePreferencesChangedNotification
inside the WebProcess. This used to happen implicitly inside WTF::platformUserPreferredLanguages() but
it is now explicit since we don't want/need it when using CFPREFS_DIRECT_MODE.

(WebKit::setPreferenceValue):
- Call CFPreferencesSetAppValue() instead of CFPreferencesSetValue(). Without this change
trying to set AppleLanguages here would have no effect and CFLocaleCopyPreferredLanguages()
would keep returning outdated languages.
- Call WTF::languageDidChange() when the AppleLanguages preference gets updated so that
  language change listeners get notified of the language change (e.g. we fire a languagechange
  event at the Window).

Source/WebKitLegacy/mac:

We need to listen for AppleLanguagePreferencesChangedNotification on mac WK1.
This used to happen implicitly inside WTF::platformUserPreferredLanguages() but
it is now explicit since we don't want/need it when using CFPREFS_DIRECT_MODE in WK2.

* WebView/WebView.mm:
(-[WebView _commonInitializationWithFrameName:groupName:]):

Source/WTF:

Update WTF::languageDidChange() to clear preferredLanguages() on Cocoa platforms
so that we get updated values from the system the next time
platformUserPreferredLanguages() is called.

platformUserPreferredLanguages() used to implicitly register a AppleLanguagePreferencesChangedNotification
listener. We've now made this registering opt-in by moving it to a separate
listenForLanguageChangeNotifications() function. This function is getting called
on Mac WK1 and WK2 when CFPREFS_DIRECT_MODE is disabled (legacy).
When CFPREFS_DIRECT_MODE is enabled in WK2, we don't want/need to listen for this
notification because the AppleLanguages preference gets pushed by the UIProcess
down to the WebProcesses. Even though we could listen for this notification,
the WebProcess would not have the latest AppleLanguages preference when receiving
the notification. This would cause us to fire the languagechange event at the
Window too early and navigator.language would keep returning the old language.
For WK2 with CFPREFS_DIRECT_MODE enabled, we now explicitly call
WTF::languageDidChange() when the "AppleLanguages" preference gets sync'd from
the UIProcess instead.

* wtf/Language.cpp:
(WTF::languageDidChange):
(WTF::platformLanguageDidChange):
* wtf/Language.h:
* wtf/cf/LanguageCF.cpp:
(WTF::languagePreferencesDidChange):
(WTF::platformLanguageDidChange):
(WTF::listenForLanguageChangeNotifications):
(WTF::platformUserPreferredLanguages):

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WebKit/OverrideAppleLanguagesPreference.mm:
(TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273851 268f45cc-cd09-0410-ab3c-d52691b4dbfc