[css-flexbox] Fix mainAxisLengthIsDefinite for orthogonal items with percentage sizes https://bugs.webkit.org/show_bug.cgi?id=222684
Reviewed by Javier Fernandez.

Source/WebCore:

r260055 fixed a crash in flexbox code by not caching the definiteness of the logical height of the flexbox
container in case of having orthogonal items because in that case, the logical heights of the flex container
and the flex item are not in the same axis. That made a lot of sense, however we should still let the method
return whether or not the child main axis is definite instead of unconditionally return false.

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::childMainSizeIsDefinite const):

LayoutTests:

* TestExpectations: Unskipped flexbox-basic-canvas-vert-001v.xhtml which is now passing.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273958 268f45cc-cd09-0410-ab3c-d52691b4dbfc