Fix potential thread safety issue in WebsiteDataStore::fetchDataAndApply() https://bugs.webkit.org/show_bug.cgi?id=222830
Reviewed by Geoffrey Garen.

The previous code was passing SecurityOriginData objects from the data store queue to the main
thread without isolated copy. This code would not be thread-safe if the function we got the
origins from would keep a copy of the origins (e.g. caching). To make this more robust from a
thread-safety standpoint, we now crossThreadCopy() to origins before passing them to the main
thread.

I also updated these to construct the WebsiteData on the data store queue instead of the main
thread for consistency with other cases inside WebsiteDataStore::fetchDataAndApply().

* UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::fetchDataAndApply):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274017 268f45cc-cd09-0410-ab3c-d52691b4dbfc