REGRESSION (r273541): Payloads requiring NSURL in serialization fail https://bugs.webkit.org/show_bug.cgi?id=222856 <rdar://problem/75036452>
Reviewed by Darin Adler.

In Bug 222145 I converted serialization of CFURLRequest to NSURLRequest. As part of that change,
I used a new encoder (WKSecureCodingURLWrapper) to wrap the NSURL to ensure proper handling of
encoded URLs.

The WKSecureCodingURLWrapper should have been declared as a subclass of NSURL, so that NSSecureCoding
would respect it as a valid way of serializing an NSURL.

This patch corrects that oversight.

* Shared/Cocoa/ArgumentCodersCocoa.mm:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274109 268f45cc-cd09-0410-ab3c-d52691b4dbfc