[GPUP] Test fast/images/animated-image-mp4.html times out when media in GPU Process is enabled https://bugs.webkit.org/show_bug.cgi?id=221795
Reviewed by Eric Carlson.

Source/WebCore:

No new tests. Fix test failures.

* platform/graphics/cocoa/IOSurface.h:
Add an empty line to make the comment clear.

* platform/graphics/cocoa/WebCoreDecompressionSession.mm:
(WebCore::WebCoreDecompressionSession::ensureDecompressionSessionForSample):
Fix an error in the attributes of a `VTDecompressionSession`.

Source/WebKit:

Pass the color space along with a `MachSendRight` created from an `IOSurface` to ensure
that we can recreate the image from the `MachSendRight`/`IOSurface` properly.

* GPUProcess/media/RemoteImageDecoderAVFProxy.cpp:
(WebKit::RemoteImageDecoderAVFProxy::createFrameImageAtIndex):
* GPUProcess/media/RemoteImageDecoderAVFProxy.h:
* GPUProcess/media/RemoteImageDecoderAVFProxy.messages.in:

* WebProcess/GPU/media/RemoteImageDecoderAVF.cpp:
(WebKit::RemoteImageDecoderAVF::createFrameImageAtIndex):
This functions needs to return a valid image when it is called in the main thread,
otherwise, the animation will pause.

LayoutTests:

* platform/wk2/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274156 268f45cc-cd09-0410-ab3c-d52691b4dbfc