Remove CSSParserContext::enforcesCSSMIMETypeInNoQuirksMode as it is unused https://bugs.webkit.org/show_bug.cgi?id=222949
Reviewed by Darin Adler.

At some point in the past we stoped doing anything with the enforcesCSSMIMETypeInNoQuirksMode
bit on CSSParserContext.

The only time it was ever disabled was when running ancient versions of iWeb, a product that
itself hasn't existed for over a decade, so I think we are ok to remove it.

Source/WebCore:

* css/parser/CSSParserContext.cpp:
(WebCore::operator==):
* css/parser/CSSParserContext.h:
(WebCore::CSSParserContextHash::hash):

Source/WebKitLegacy/mac:

* WebView/WebPreferencesDefaultValues.h:
* WebView/WebPreferencesDefaultValues.mm:
(WebKit::defaultEnforceCSSMIMETypeInNoQuirksMode): Deleted.

Source/WTF:

* Scripts/Preferences/WebPreferences.yaml:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274168 268f45cc-cd09-0410-ab3c-d52691b4dbfc