Clean up the text boxes when the associated text renderer is detached from the tree https://bugs.webkit.org/show_bug.cgi?id=223009 <rdar://74788950>
Reviewed by Antti Koivisto.

This is the case when the text renderer gets moved under a different block and the original block
becomes non-legacy based line layout.
The inline text boxes are normally cleaned up either when the associated renderer is destroyed
or during the subsequent layout when the legacy code invalidates the line.
However when we switch from legacy to modern line layout this invalidation never runs.
This patch ensures that the text boxes are destroyed soon after the renderer is detached.

* rendering/updating/RenderTreeBuilder.cpp:
(WebCore::RenderTreeBuilder::detachFromRenderElement):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274211 268f45cc-cd09-0410-ab3c-d52691b4dbfc