Improve font-variation-settings interpolation https://bugs.webkit.org/show_bug.cgi?id=223027
Reviewed by Dean Jackson.

LayoutTests/imported/w3c:

Add an extra 140 PASS results.

* web-platform-tests/css/css-fonts/inheritance-expected.txt:
* web-platform-tests/css/css-fonts/animations/font-variation-settings-composition-expected.txt:
* web-platform-tests/css/css-fonts/animations/font-variation-settings-interpolation-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-001-expected.txt:

Source/WebCore:

The animation wrapper for font-variation-settings had no canInterpolate() override. We now
implement such a method with the same logic used for blendFunc() method for FontVariationSettings
which only blends when the values have the same number of entries, and the same tag for each
entry.

We also modify the blendFunc() method to remove the return of empty values since it should only
be called in a scenario where the method can blend the values (due to canInterpolate() being
implemented) or with progress equal to 0 or 1 if the animation is discrete. In the latter cases,
we simply return the "from" or "to" values as-is.

Finally, to pass the tests for the "initial" and "inherit" values, we implement the custom
style builder methods for font-variation-settings.

* animation/CSSPropertyAnimation.cpp:
(WebCore::blendFunc):
* style/StyleBuilderCustom.h:
(WebCore::Style::BuilderCustom::applyInitialFontVariationSettings):
(WebCore::Style::BuilderCustom::applyInheritFontVariationSettings):

LayoutTests:

Remove an incorrect test that is also testing functionality well covered by the WPT test
css/css-fonts/animations/font-variation-settings-interpolation.html.

* animations/font-variations/font-variation-settings-unlike-expected.txt: Deleted.
* animations/font-variations/font-variation-settings-unlike.html: Deleted.
* platform/win/TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274235 268f45cc-cd09-0410-ab3c-d52691b4dbfc