REGRESSION(r272293) WebArchives originally loaded over HTTP fail to load subresources that would be upgraded to HTTPS https://bugs.webkit.org/show_bug.cgi?id=223044 <rdar://75228599>
Patch by Alex Christensen <achristensen@webkit.org> on 2021-03-11
Reviewed by Chris Dumez.

Source/WebCore:

Existing WebArchive content has keys that are URLs that have the "http" scheme.
To continue to successfully load these WebArchives, if we can't find a resource and we tried an HTTPS URL,
also try the HTTP URL because it may have been upgraded from HTTP to HTTPS in the loader.
We aren't going to the network anyways, so there is no advantage in pretending to use HTTPS.

* loader/archive/ArchiveResourceCollection.cpp:
(WebCore::ArchiveResourceCollection::archiveResourceForURL):

Tools:

* TestWebKitAPI/Tests/mac/LoadWebArchive.mm:
(TestWebKitAPI::TEST):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274283 268f45cc-cd09-0410-ab3c-d52691b4dbfc