Fix interpolation of perspective property https://bugs.webkit.org/show_bug.cgi?id=223111
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Add an extra 59 PASS results, although we also have 18 new FAIL results. Of those
new failures, 16 are due to what I believe to be an issue in the WPT test and I filed
https://github.com/web-platform-tests/wpt/pull/28036 to address this. With this change
in the test, we pass all perspective interpolation tests.

* web-platform-tests/css/css-transforms/animation/perspective-composition-expected.txt:
* web-platform-tests/css/css-transforms/animation/perspective-interpolation-expected.txt:

Source/WebCore:

In order to correctly interplate the "perspective" CSS property, we must not interpolate
between "none" values and lengths. To do this, we add a new wrapper for this property with
a canInterpolate() implementation that uses RenderStyle::hasPerspective() to determine
whether we're dealing with a "none" value.

We also had to make a change to the way the "none" value is represented internally, since it
used to be 0 although the spec (https://drafts.csswg.org/css-transforms-2/#perspective-property)
says "perspective: 0 in a stylesheet will still serialize back as 0". So we now change the
initial value to be -1, which is fine since negative values are otherwise not allowed.

To correctly support this, we must also change consumePerspective() to no longer disallow the
0 value at parse time.

* animation/CSSPropertyAnimation.cpp:
(WebCore::PerspectiveWrapper::PerspectiveWrapper):
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):
* css/parser/CSSPropertyParser.cpp:
(WebCore::consumePerspective):
* rendering/style/RenderStyle.h:
(WebCore::RenderStyle::hasPerspective const):
(WebCore::RenderStyle::initialPerspective):
* style/StyleBuilderConverter.h:
(WebCore::Style::BuilderConverter::convertPerspective):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274329 268f45cc-cd09-0410-ab3c-d52691b4dbfc