Support animation of perspective-origin property https://bugs.webkit.org/show_bug.cgi?id=223116
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Add an extra 35 PASS results.

* web-platform-tests/css/css-transforms/animation/perspective-origin-interpolation-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/accumulation-per-property-002-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/addition-per-property-002-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt:

Source/WebCore:

While we have support for animating "perspective-origin-x" and "perspective-origin-y", which are
not part of the CSS Transforms standard, we do not support animation of "perspective-origin",
which we consider in WebKit to be a shorthand property. All that is needed to address this is to
add CSSPropertyPerspectiveOrigin in the list of animatable shorthand properties when creating
animation wrappers.

* animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274355 268f45cc-cd09-0410-ab3c-d52691b4dbfc