[selectors] :focus-visible matches body after keyboard event https://bugs.webkit.org/show_bug.cgi?id=223113
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

* web-platform-tests/css/selectors/focus-visible-001.html: Modify the test to verify
that only the element matches :focus-visible (and not the body).
* web-platform-tests/css/selectors/focus-visible-019-expected.txt: Added.
* web-platform-tests/css/selectors/focus-visible-019.html: Added new test to check script focus in keyboard event,
and that again only the element matches :focus-visible (and not the body).

Source/WebCore:

Fix the bug with some changes in EventHandler::internalKeyEvent().
When you use TAB (or other key) the |element| variable in this method is the document body,
however that element is not focused (element->focused() is false).
Before this patch we were marking the element as matchin :focus-visible,
however we shouldn't do that if the element is not focused (added a condition to avoid doing that).

Apart from that this patch also fixes a related issue, if a keyboard event handler is changing focus via script
there's a part of this method that takes care of updating the |element| variable.
In that case we have to remove :focus-visible flag from the previous element, and add it to the new one.

Test: imported/w3c/web-platform-tests/css/selectors/focus-visible-001.html
Test: imported/w3c/web-platform-tests/css/selectors/focus-visible-019.html

* page/EventHandler.cpp:
(WebCore::EventHandler::internalKeyEvent):

LayoutTests:

* platform/ios/TestExpectations: Skip new test for iOS.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274365 268f45cc-cd09-0410-ab3c-d52691b4dbfc