[iOS] Selecting the first word in an image overlay may select text in the previous line https://bugs.webkit.org/show_bug.cgi?id=223153
Reviewed by Tim Horton.

Source/WebCore:

Add a (collapsible) newline at the start of each line of text in an image overlay, so that
`wordRangeFromPosition` will not include content from the previous line when selecting the first word in a
line inside an image overlay.

Test: fast/images/image-extraction/ios/select-word-in-image-overlay.html

* html/HTMLElement.cpp:
(WebCore::HTMLElement::updateWithImageExtractionResult):

LayoutTests:

Add a layout test to verify the behavior change.

* fast/images/image-extraction/ios/select-word-in-image-overlay-expected.txt: Added.
* fast/images/image-extraction/ios/select-word-in-image-overlay.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274396 268f45cc-cd09-0410-ab3c-d52691b4dbfc