Stop calling [NSHTTPCookieStorage sharedHTTPCookieStorage] in the WebProcess https://bugs.webkit.org/show_bug.cgi?id=223186 <rdar://75018105>
Reviewed by Geoffrey Garen.

NetworkStorageSession::nsCookieStorage() was calling [NSHTTPCookieStorage sharedHTTPCookieStorage]
even if cookieStorage() is not nil. I updated the condition so that we do not attempt to call
[NSHTTPCookieStorage sharedHTTPCookieStorage] when the m_isInMemoryCookieStore flag is set. The
WebProcess is using an in-memory cookie store (thus has m_isInMemoryCookieStore flag set) and
is definitely not using the shared HTTP cookie store.

* platform/network/cocoa/NetworkStorageSessionCocoa.mm:
(WebCore::NetworkStorageSession::nsCookieStorage const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274429 268f45cc-cd09-0410-ab3c-d52691b4dbfc